Move non instance methods to modules
