+++ app/models/pack_record.rb
@@ -3,7 +3,6 @@ class PackRecord < ActiveRecord::Base
   enum status: [:DISPATCHED, :RECEIVED, :COMPLETED]
   scope :overdue, -> { where("due_date < ?", Date.today ) }
   store_accessor :comments, :speed, :accuracy, :completion, :presentation
-#  validate_presence_of :comments if self.status == 'RECEIVED'
 
   def calculate_reward(user, score, pack_record)
     if score >= 70 && score <= 79
