+++ lib/validator_mod.rb
@@ -1,0 +1,14 @@
+module ValidatorMod
+  extend self
+
+  def validate_rewards(user)
+    if Date.today == user.activation_date + 6.months
+      user.rewards = 0.00
+      user.activation_date = Date.today
+      user.save 
+    elsif user.activation_date + 5.months == Date.today
+      UserMailer.reward_expiry_reminder(user).deliver_now
+    end 
+  end
+
+end
